input: cyttsp-i2c-qc: change code path for cyttsp_xy_handler
cyttsp_xy_handler is the function which sends the coordinates
to input framework. In the starting of the function it checks
the mode of the controller. But while going to suspend,
this check is not required, since the controller will
eventually goes to power collapse. So this change will
skip the check if it is going to suspend.

Change-Id: If66a8f0bb55fad42a375810ee75876ac43e34511
Signed-off-by: Shantanu Jain <shjain@codeaurora.org>