prima: Reduce expiration time of traffic timer during roaming
Currently, traffic timer started for 8secs after successful connection to
allow for key completion and DHCP. As there is no DHCP during roaming,
this delay is not required. So, stop this timer during roaming and start a
timer of minimal duration to allow STA to enter power save immediately
after key completion.

Change-Id: I6c6111fc1aacb524da463914d357710a7f125f4a
CRs-Fixed: 868916